add pokedex, trainer bios, loss condition.

+++ src/components/Navbar.js
@@ -3,21 +3,26 @@ import { useHistory } from "react-router-dom";
 import { AppContext } from '../context/AppContext'
 
 const Navbar = () => {
+    const { battleSequence } = useContext(AppContext)
     let history = useHistory()
 
     const routeToHome = () => {
+        if(battleSequence) return
         history.push('../home')
     }
 
     const routeToPokedex = () => {
+        if (battleSequence) return
         history.push('../pokedex')
     }
 
     const routeToTrainerBio = () => {
-        history.push('../trainers')
+        if (battleSequence) return
+        history.push('../selectbio/gymleaders')
     }
 
     const signout = () => {
+        if (battleSequence) return
         localStorage.clear()
         history.push('../')
     }
@@ -26,7 +31,7 @@ const Navbar = () => {
         <div className="navbar">
             <div className="tab" onClick={routeToHome}><p>Home</p></div>
             <div className="tab" onClick={routeToPokedex}><p>Pokedex</p></div>
-            {/* <div className="tab" onClick={routeToTrainerBio}><p>Trainer Bio</p></div> */}
+            <div className="tab" onClick={routeToTrainerBio}><p>Bio</p></div>
             <div className='tab signout-tab' onClick={signout}><p>Sign Out</p></div>
         </div>
     )
